kimi_k3: the EP declaration tests move into the existing K3 test file
Review ask: the standalone file goes. The two checks keep their value -- the
sharding config is set on the config tree and constructing configs directly
has dropped declarations invisibly before, so plain-dp-leaves-the-tree-alone
and EP-shards-the-expert-axis are worth pinning without a GPU -- but they live
in tests/unit_tests/test_kimi_k3.py now, the file the original K3 PR created,
rather than establishing a per-model sharding-test file this repo does not
otherwise have.
